compiler: use qualified Names for definitions

+++ compiler/test_flatten_classes.py
@@ -21,7 +21,7 @@ from utils_test import MockPackageLoader
 
 class TestFlattenClasses(unittest.TestCase):
     def setUp(self):
-        self.rootMethodNames = [m.name for m in getRootClass().methods]
+        self.rootMethodNames = [m.name.short() for m in getRootClass().methods]
 
     def analyzeFromSource(self, source):
         filename = "(test)"
@@ -41,7 +41,7 @@ class TestFlattenClasses(unittest.TestCase):
     def testSimpleClass(self):
         info = self.analyzeFromSource("class C")
         C = info.package.findClass(name="C")
-        self.assertEquals(self.rootMethodNames, [m.name for m in C.methods])
+        self.assertEquals(self.rootMethodNames, [m.name.short() for m in C.methods])
 
     def testDerivedClassWithMethods(self):
         source = "class A\n" + \
@@ -50,7 +50,8 @@ class TestFlattenClasses(unittest.TestCase):
                  "  def g = 12\n"
         info = self.analyzeFromSource(source)
         B = info.package.findClass(name="B")
-        self.assertEquals(self.rootMethodNames + ["f", "g"], [m.name for m in B.methods])
+        self.assertEquals(self.rootMethodNames + ["f", "g"],
+                          [m.name.short() for m in B.methods])
 
     def testDerivedClassWithOverloads(self):
         source = "class A\n" + \
@@ -59,7 +60,8 @@ class TestFlattenClasses(unittest.TestCase):
                  "  def f(x: i64) = x\n"
         info = self.analyzeFromSource(source)
         B = info.package.findClass(name="B")
-        self.assertEquals(self.rootMethodNames + ["f", "f"], [m.name for m in B.methods])
+        self.assertEquals(self.rootMethodNames + ["f", "f"],
+                          [m.name.short() for m in B.methods])
 
     def testDerivedClassWithOverride(self):
         source = "class A\n" + \
@@ -68,7 +70,8 @@ class TestFlattenClasses(unittest.TestCase):
                  "  def f(x: i32) = 12i32\n"
         info = self.analyzeFromSource(source)
         B = info.package.findClass(name="B")
-        self.assertEquals(self.rootMethodNames + ["f"], [m.name for m in B.methods])
+        self.assertEquals(self.rootMethodNames + ["f"],
+                          [m.name.short() for m in B.methods])
 
     def testMustOverrideInheritedAbstractMethod(self):
         source = "abstract class A\n" + \
